Update start_rust.sh
A few tweeks

+++ start_rust.sh
@@ -53,7 +53,15 @@ trap 'exit_handler' SIGHUP SIGINT SIGQUIT SIGTERM
 ##################################
 
 compilerdir="/steamcmd/rust"
-find "${compilerdir:?}" -type f -name "Compiler.x86_x64" -delete
+
+if [ -f "/steamcmd/rust/Compiler.x86_x64" ]; then
+  find "${compilerdir:?}" -type f -name "Compiler.x86_x64" -delete
+ 2>&1 /dev/null
+  elif [ -f "/steamcmd/rust/CSharpCompiler.x86_x64" ]; then
+  find "${compilerdir:?}" -type f -name "CSharpCompiler.x86_x64" -delete
+ 2>&1 /dev/null
+  fi
+fi
 
 
 #DEFAULT VARIABLES.
